Fix Sonar duplicate code warning
